feat: simple implementation of ECS

+++ include/ECS/Components/PositionComponent.hpp
@@ -4,11 +4,11 @@
 #include "Vector2.hpp"
 
 namespace sex {
-    class PositionComponent : public Vector2<float>, public IComponent {
+    class PositionComponent : public Vector2<double>, public IComponent {
         public:
-            PositionComponent(float X = 0, float Y = 0) : Vector2<float>(X, Y) {}
-            PositionComponent(Vector2 const &v) : Vector2<float>(v) {}
-            PositionComponent(Vector2 const &&v) : Vector2<float>(v) {}
+            PositionComponent(double X = 0, double Y = 0) : Vector2<double>(X, Y) {}
+            PositionComponent(Vector2 const &v) : Vector2<double>(v) {}
+            PositionComponent(Vector2 const &&v) : Vector2<double>(v) {}
 
             ~PositionComponent() {}
     };
